refactor: extract story feature into StoryReport(Report) in story.py
Move the AI story-annotation tab off the base Report onto a StoryReport
subclass so the base report stays free of story state and UI. Reports in
the app now instantiate StoryReport, base Report keeps only the core
report/export behaviour.
