+++ lumen/ai/report.py
@@ -24,23 +24,20 @@ from panel.viewable import Viewable, Viewer
 from panel_material_ui import (
     Accordion, BreakpointSwitcher, Button, Card, ChatFeed, ChatMessage,
     Checkbox, Container, Dialog, Divider, FileDownload, IconButton, MenuButton,
-    Progress, Select, SpeedDial, Tabs, TextAreaInput, TextInput, Typography,
+    Progress, Select, SpeedDial, TextAreaInput, TextInput, Typography,
 )
 
-from ..config import config
-from ..pipeline import Pipeline
 from ..views.base import Panel, View
 from .actor import (
     Actor, ContextProvider, NullStep, TContext,
 )
-from .agents.story import StoryAgent, StoryState, build_catalog
 from .config import MissingContextError
 from .context import (
     LWW, ContextError, ValidationIssue, collect_task_outputs,
     input_dependency_keys, merge_contexts, validate_task_inputs,
     validate_taskgroup_exclusions,
 )
-from .editors import EditableProse, LumenEditor
+from .editors import LumenEditor
 from .export import (
     format_output, make_md_cell, make_preamble, write_notebook,
 )
@@ -890,15 +887,6 @@ class Report(TaskGroup):
 
     level = 1
 
-    # The Story tab is appended after the Report tab once a story exists.
-    _story_tab = 1
-
-    _tone_presets = [
-        ("Executive summary", "Write a concise executive summary for decision-makers, leading with the key takeaways."),
-        ("Technical deep-dive", "Write a detailed technical analysis that explains the methods and the data behind each chart."),
-        ("Highlights only", "Keep it brief: just the headline finding for each chart and table."),
-    ]
-
     def __init__(self, *tasks, **params):
         if not tasks:
             tasks = params.pop('tasks', [])
@@ -936,7 +924,6 @@ class Report(TaskGroup):
 
     def _init_view(self):
         self._section_headers = {}
-        self._story = StoryState()
         self._header_title = Typography(
             self.param.title, variant="h1", margin=(0, 0, 0, 10)
         )
@@ -948,22 +935,6 @@ class Report(TaskGroup):
                     "overflowX": "auto"
                 }
             }
-        )
-        # The story lives in its own tab so the section accordion (and its
-        # checkboxes) is never disturbed, and the two are told apart at a glance
-        # rather than by swapping one view for the other. A headerless Card gives
-        # it the same white pane as the section cards.
-        self._story_column = Card(
-            sizing_mode="stretch_width", margin=(0, 5, 15, 5),
-            collapsible=False, hide_header=True,
-        )
-        # The Story tab is added only once a story exists.
-        # Let the tab panel grow with its content (the report/story scrolls at
-        # the container level, below) instead of the tab clipping a tall body.
-        self._tabs = Tabs(
-            ("Report", self._view),
-            sizing_mode="stretch_width",
-            stylesheets=[".MuiTabsPanel > .MuiBox-root { overflow: visible; }"],
         )
         self._run = IconButton(
             icon="play_arrow", on_click=self._execute_event, margin=0, size="large",
@@ -987,16 +958,6 @@ class Report(TaskGroup):
         self._settings = IconButton(
             icon="settings", on_click=self._open_settings, size="large", color="default",
             margin=0, description="Configure Report", visible=False
-        )
-        self._annotate = IconButton(
-            icon="auto_stories", on_click=self._open_story_dialog, size="large",
-            color="default", margin=0, visible=False,
-            description="Annotate the report with an AI-written story",
-        )
-        self._arrange = IconButton(
-            icon="reorder", on_click=self._open_arrange_dialog, size="large",
-            color="default", margin=0, visible=False,
-            description="Arrange the report (reorder sections and add headings)",
         )
         self._export = MenuButton(
             label="", icon="get_app", variant="text", color="default",
@@ -1032,72 +993,12 @@ class Report(TaskGroup):
             close_on_click=True,
             title=f"Report Settings: {self.title}",
         )
-        self._outline_editor = pn.widgets.JSONEditor(
-            mode="tree", sizing_mode="stretch_width", height=400,
-        )
-        self._outline_editor.param.watch(self._on_outline_change, "value")
-        self._outline_apply = Button(
-            label="Apply", variant="contained", on_click=self._apply_outline,
-        )
-        self._outline_dialog = Dialog(
-            self._outline_editor,
-            Row(self._outline_apply, align="end", margin=(10, 0, 0, 0), sizing_mode="stretch_width"),
-            show_close_button=True,
-            title="Arrange Report",
-            width_option="md",
-        )
-        self._story_guidance = TextAreaInput(
-            placeholder="Guide the story, e.g. focus on seasonal trends (optional)",
-            sizing_mode="stretch_width", rows=3, margin=(10, 0, 0, 0),
-        )
-        self._story_presets = Row(
-            *[
-                Button(
-                    label=label, variant="outlined", size="small", color="default",
-                    on_click=partial(self._apply_story_preset, text=text),
-                )
-                for label, text in self._tone_presets
-            ],
-            sizing_mode="stretch_width",
-        )
-        self._story_generate = Button(
-            label="Generate", variant="contained", icon="auto_stories",
-            on_click=self._generate_story,
-        )
-        self._story_dialog = Dialog(
-            Typography("Pick a tone or write your own guidance, then generate the story.", variant="body2"),
-            self._story_presets,
-            self._story_guidance,
-            Row(self._story_generate, align="end", margin=(10, 0, 0, 0), sizing_mode="stretch_width"),
-            show_close_button=True,
-            title="Annotate Report",
-            width_option="sm",
-        )
-        # Rewriting a single paragraph with the LLM.
-        self._edit_index = None
-        self._edit_instruction = TextAreaInput(
-            placeholder="How should this paragraph change? e.g. make it shorter and lead with the number",
-            sizing_mode="stretch_width", rows=3, margin=(10, 0, 0, 0),
-        )
-        self._edit_apply = Button(
-            label="Rewrite", variant="contained", icon="auto_fix_high",
-            on_click=self._rewrite_prose,
-        )
-        self._edit_dialog = Dialog(
-            self._edit_instruction,
-            Row(self._edit_apply, align="end", margin=(10, 0, 0, 0), sizing_mode="stretch_width"),
-            show_close_button=True,
-            title="Edit paragraph with AI",
-            width_option="sm",
-        )
         self._menu = Row(
             self._header_title,
             self._run,
             self._stop,
             self._clear,
             self._collapse,
-            self._annotate,
-            self._arrange,
             self._export,
             self._settings,
             sizing_mode="stretch_width"
@@ -1107,8 +1008,6 @@ class Report(TaskGroup):
                 {"label": "Execute Report", "icon": "play_arrow"},
                 {"label": "Stop Report", "icon": "stop"},
                 {"label": "Clear Report", "icon": "clear"},
-                {"label": "Annotate Report", "icon": "auto_stories"},
-                {"label": "Arrange Report", "icon": "reorder"},
                 {"label": "Export as Notebook", "icon": "description", "format": "ipynb"},
                 {"label": "Export as HTML", "icon": "language", "format": "html"},
                 {"label": "Configure Report", "icon": "settings"}
@@ -1136,21 +1035,12 @@ class Report(TaskGroup):
             "Untick a chart, table or note to leave it out of the exported report.",
             variant="body2", margin=(0, 10, 10, 10),
             sx={"color": "text.secondary"},
-            # Only on the report tab: the story tab has nothing to untick.
-            visible=param.bind(
-                lambda status, active: (
-                    status in ("success", "error", "cancelled") and active == 0
-                ),
-                self.param.status, self._tabs.param.active,
-            ),
+            visible=self.param.status.rx().rx.in_(("success", "error", "cancelled")),
         )
         self._container = Column(
             self._export_hint,
-            self._tabs,
+            self._view,
             self._dialog,
-            self._outline_dialog,
-            self._story_dialog,
-            self._edit_dialog,
             margin=(0, 0, 0, 5),
             sizing_mode="stretch_both",
             height_policy='fit',
@@ -1169,10 +1059,6 @@ class Report(TaskGroup):
             self._handle_cancel()
         elif icon == "clear":
             self.reset()
-        elif icon == "auto_stories":
-            self._open_story_dialog()
-        elif icon == "reorder":
-            self._open_arrange_dialog()
         elif icon in ("description", "language"):
             fmt = item.get("format", "ipynb")
             self._export.value = {"format": fmt}
@@ -1191,9 +1077,6 @@ class Report(TaskGroup):
         has_outputs = self.status in ("success", "error", "cancelled") or bool(self.views)
         self._clear.visible = has_outputs
         self._collapse.visible = has_outputs
-        self._annotate.visible = has_outputs
-        self._annotate.disabled = self.llm is None
-        self._arrange.visible = has_outputs
         self._export.visible = has_outputs
         self._settings.visible = has_outputs
         # Only animate play button when no outputs
@@ -1212,8 +1095,6 @@ class Report(TaskGroup):
         if self._active_task is not None and not self._active_task.done():
             # Already running; the stop button handles cancellation
             return
-        # A fresh run invalidates any previously generated story.
-        self._discard_story()
         await asyncio.sleep(0.01)  # yield the event loop to allow button loading state to update
         task = asyncio.create_task(self.execute())
         self._active_task = task
@@ -1236,24 +1117,8 @@ class Report(TaskGroup):
         if not len(self):
             return self.views
         views = list(self._header)
-        # When the story view is shown, export the blog-post flow: the story
-        # title, then prose interleaved with the charts and tables it discusses.
-        if self._show_story and self._story.blocks:
-            views.append(Typography(self._story.title, variant="h2"))
-            views += [
-                Markdown(obj) if kind == "prose" else obj
-                for kind, obj in self._story.blocks
-            ]
-            return views
-        if self._story.outline:
-            for kind, value, level in self._iter_arrangement():
-                if kind == "heading":
-                    views.append(Typography(value, variant=f"h{level}"))
-                else:
-                    self._append_section_views(views, value)
-        else:
-            for section in self:
-                self._append_section_views(views, section)
+        for section in self:
+            self._append_section_views(views, section)
         return views
 
     def _append_section_views(self, views, section):
@@ -1274,46 +1139,20 @@ class Report(TaskGroup):
             views += self._selected_views(task) if isinstance(task, TaskGroup) else list(task.views)
         return views
 
-    def _section_by_title(self, title):
-        for section in self:
-            if section.title == title:
-                return section
-        return None
-
-    def _build_default_outline(self):
-        """Seed the arrangement outline from the currently selected sections."""
-        return [{"section": section.title} for section in self if section.include_in_export]
-
     def _export_view(self):
         # Task-less reports (e.g. Report.from_views) populate ``_view`` directly.
         if not len(self):
             return self._view
-        # When the story view is shown, save the same blog-post flow shown live:
-        # prose interleaved with the charts and tables, matching the notebook.
-        if self._show_story and self._story.blocks:
-            return Column(*self._story_flow(), sizing_mode="stretch_width", margin=(0, 5, 15, 5))
         # Build each card from the section's selected views rather than the
         # section itself, so a discarded nested section is left out of the HTML
         # too. The card header already shows the title, hence no heading.
-        cards = []
-        if self._story.outline:
-            # Follow the arranged outline so HTML matches the notebook order.
-            for kind, value, _ in self._iter_arrangement():
-                if kind == "heading":
-                    cards.append((value, Markdown(f"### {value}")))
-                else:
-                    cards.append((
-                        value.title,
-                        Column(*self._selected_views(value, include_header=False), sizing_mode="stretch_width"),
-                    ))
-        else:
-            cards += [
-                (
-                    section.title,
-                    Column(*self._selected_views(section, include_header=False), sizing_mode="stretch_width"),
-                )
-                for section in self if section.include_in_export
-            ]
+        cards = [
+            (
+                section.title,
+                Column(*self._selected_views(section, include_header=False), sizing_mode="stretch_width"),
+            )
+            for section in self if section.include_in_export
+        ]
         return Accordion(
             *cards,
             active=list(range(len(cards))),
@@ -1400,261 +1239,14 @@ class Report(TaskGroup):
     def _open_settings(self, event=None):
         self._dialog.open = True
 
-    def _open_arrange_dialog(self, event=None):
-        # Seed from the current arrangement, or the default section order.
-        self._outline_editor.value = self._story.outline or self._build_default_outline()
-        self._outline_dialog.open = True
-
-    def _on_outline_change(self, event):
-        self._story.outline = event.new or []
-        self._populate_view()
-
-    def _apply_outline(self, event=None):
-        # Read the editor's current value explicitly and apply the arrangement
-        # to both the live report and the exports.
-        self._story.outline = list(self._outline_editor.value or [])
-        self._populate_view()
-        self._outline_dialog.open = False
-
-    def _open_story_dialog(self, event=None):
-        self._story_dialog.open = True
-
-    def _apply_story_preset(self, event=None, text=""):
-        self._story_guidance.value = text
-
-    async def _generate_story(self, event=None):
-        guidance = (self._story_guidance.value or "").strip()
-        self._story_dialog.open = False
-        await self._annotate_report(guidance=guidance)
-
-    def _collect_story_views(self):
-        """Charts and tables from the selected sections, in reading order."""
-        return [
-            view
-            for section in self if section.include_in_export
-            for view in section.views if isinstance(view, LumenEditor)
-        ]
-
-    async def _annotate_report(self, event=None, guidance=""):
-        """Write an AI blog-post story over the selected charts/tables and render it."""
-        if self.llm is None or not len(self):
-            return
-        catalog = self._collect_story_views()
-        if not catalog:
-            return
-        self._annotate.loading = True
-        self._story_generate.loading = True
-        try:
-            agent = StoryAgent(llm=self.llm)
-            story = await agent.write_story(
-                await build_catalog(catalog), guidance=guidance, title=self.title or "",
-            )
-            self._set_story(story, catalog)
-        except Exception:
-            tb.print_exc()
-        finally:
-            self._annotate.loading = False
-            self._story_generate.loading = False
-
-    def _set_story(self, story, catalog):
-        """Resolve the LLM story into interleaved prose/view blocks as a new version."""
-        blocks, used = [], set()
-        for block in story.blocks:
-            if block.prose and block.prose.strip():
-                blocks.append(("prose", block.prose))
-            index = block.view
-            if index is not None and 1 <= index <= len(catalog) and index not in used:
-                used.add(index)
-                blocks.append(("view", catalog[index - 1]))
-        # Never drop a selected chart/table: append any the model did not place.
-        for index, view in enumerate(catalog, start=1):
-            if index not in used:
-                blocks.append(("view", view))
-        # Each story is kept as its own version, so regenerating offers a fresh
-        # take without throwing away one the user may have edited by hand.
-        self._story.versions.append({"title": story.title or "Story", "blocks": blocks})
-        self._select_version(len(self._story.versions) - 1)
-
-    def _select_version(self, index):
-        """Show one of the generated versions of the story."""
-        self._story.version = index
-        version = self._story.versions[index]
-        self._story.title = version["title"]
-        # The version's own list is used live, so edits stay with that version.
-        self._story.blocks = version["blocks"]
-        self._render_story()
-
-    def _switch_version(self, event):
-        self._select_version(event.new)
-
-    def _rebuild_view(self, editor):
-        """An independent copy of a chart/table so the story preview never detaches
-        it from its section; data is embedded so the copy stands alone."""
-        component = editor.component
-        with config.param.update(serializer='csv'):
-            spec = component.to_spec()
-        return Pipeline.from_spec(spec) if isinstance(component, Pipeline) else View.from_spec(spec)
-
-    def _story_flow(self, editable=False):
-        """
-        The blog-post flow as renderables: the title, then prose interleaved with
-        freshly rebuilt charts and tables. Rebuilding gives each caller its own
-        copies so the live view and the exports never share objects. Prose is
-        editable on screen and plain Markdown when exported.
-        """
-        items = []
-        if self._story.title:
-            heading = Typography(self._story.title, variant="h4", margin=(10, 10, 0, 10))
-            if editable and self.llm is not None:
-                controls = [heading]
-                if len(self._story.versions) > 1:
-                    version = Select(
-                        options={f"Version {i+1}": i for i in range(len(self._story.versions))},
-                        value=self._story.version, width=140, align="center",
-                        margin=(10, 10, 0, 0), label="",
-                        description="Switch between the generated versions of the story",
-                    )
-                    version.param.watch(self._switch_version, "value")
-                    controls.append(version)
-                # Regenerating reopens the guidance popup, so a fresh story can be
-                # written with a different angle as a new version.
-                controls.append(Button(
-                    label="Regenerate", variant="outlined", size="small", color="default",
-                    icon="refresh", align="center", margin=(10, 10, 0, 0),
-                    description="Write the story again as a new version",
-                    on_click=self._open_story_dialog,
-                ))
-                heading = Row(*controls, align="center", sizing_mode="stretch_width")
-            items.append(heading)
-        for index, (kind, obj) in enumerate(self._story.blocks):
-            if kind != "prose":
-                items.append(self._story_figure(obj))
-            elif editable:
-                items.append(self._editable_prose(index, obj))
-            else:
-                items.append(Markdown(obj, sizing_mode="stretch_width", margin=(5, 10)))
-        return items
-
-    def _story_figure(self, editor):
-        """A chart or table as a fixed-height figure, so it flows in the story
-        like a blog-post image instead of stretching to fill and swallowing the
-        scroll of the tab it lives in."""
-        return Column(
-            self._rebuild_view(editor),
-            height=400, sizing_mode="stretch_width", margin=(5, 10),
-            styles={"min-height": "unset"},
-        )
-
-    def _editable_prose(self, index, text):
-        """Prose the user can edit in place, by hand or by asking the LLM."""
-        prose = EditableProse(value=text, sizing_mode="stretch_width", margin=(5, 10))
-        prose.param.watch(partial(self._update_prose, index), "value")
-        edit = IconButton(
-            icon="auto_fix_high", size="small", color="default", align="start",
-            margin=(5, 0, 0, 0), visible=self.llm is not None,
-            description="Rewrite this paragraph with AI",
-            on_click=partial(self._open_edit_dialog, index),
-        )
-        return Row(prose, edit, sizing_mode="stretch_width")
-
-    def _update_prose(self, index, event):
-        # Keep the story blocks as the single source of truth so manual edits
-        # survive tab switches and flow into every export.
-        self._story.blocks[index] = ("prose", event.new)
-
-    def _open_edit_dialog(self, index, event=None):
-        self._edit_index = index
-        self._edit_instruction.value = ""
-        self._edit_dialog.open = True
-
-    async def _rewrite_prose(self, event=None):
-        """Ask the LLM to rewrite the paragraph the edit dialog was opened on."""
-        instruction = (self._edit_instruction.value or "").strip()
-        if self.llm is None or self._edit_index is None or not instruction:
-            return
-        self._edit_dialog.open = False
-        self._edit_apply.loading = True
-        try:
-            agent = StoryAgent(llm=self.llm)
-            edit = await agent.rewrite_prose(
-                self._story.blocks[self._edit_index][1],
-                instruction,
-                await build_catalog(self._collect_story_views()),
-            )
-            self._story.blocks[self._edit_index] = ("prose", edit.prose)
-            self._render_story()
-        except Exception:
-            tb.print_exc()
-        finally:
-            self._edit_apply.loading = False
-
-    @property
-    def _show_story(self):
-        """Whether the story, rather than the report, is the view on screen."""
-        return len(self._tabs) > self._story_tab and self._tabs.active == self._story_tab
-
-    def _render_story(self):
-        """Build the blog-post flow into the Story tab, then switch to it."""
-        self._story_column[:] = self._story_flow(editable=True)
-        if len(self._tabs) <= self._story_tab:
-            self._tabs.append(("Story", self._story_column))
-        self._tabs.active = self._story_tab
-
-    def _discard_story(self):
-        """Forget every generated version and drop back to the report tab."""
-        self._story.blocks = []
-        self._story.title = ""
-        self._story.versions = []
-        self._story.version = 0
-        self._story_column[:] = []
-        if len(self._tabs) > self._story_tab:
-            self._tabs.pop(self._story_tab)
-        # Popping leaves ``active`` pointing at the removed tab.
-        self._tabs.active = 0
-
-    def reset(self, start: int = 0):
-        # Clearing the report also drops the generated story.
-        self._discard_story()
-        super().reset(start)
-
     def _section_header(self, section):
         """Card header with a checkbox to keep/discard the section on export."""
         return _export_header(section, self)
 
-    def _iter_arrangement(self, selected_only=True):
-        """
-        Yield the arranged outline as ('heading', text, level) or
-        ('section', section, None) items, resolving section titles and
-        dropping duplicates (and, by default, discarded sections).
-        """
-        seen = set()
-        for block in self._story.outline:
-            if "heading" in block:
-                yield "heading", block["heading"], block.get("level", 2)
-            elif "section" in block:
-                section = self._section_by_title(block["section"])
-                if section is None or id(section) in seen:
-                    continue
-                if selected_only and not section.include_in_export:
-                    continue
-                seen.add(id(section))
-                yield "section", section, None
-
-    def _ordered_sections(self):
-        """Sections in the arranged outline order (unreferenced ones kept at the end)."""
-        if not self._story.outline:
-            return list(self)
-        ordered = [s for kind, s, _ in self._iter_arrangement(selected_only=False) if kind == "section"]
-        seen = {id(section) for section in ordered}
-        for section in self:
-            if id(section) not in seen:
-                ordered.append(section)
-        return ordered
-
     def _populate_view(self):
         headers = {}
         objects = []
-        for section in self._ordered_sections():
+        for section in self:
             header = self._section_headers.get(section) or self._section_header(section)
             headers[section] = header
             objects.append((header, section))
